Rename with_vector to with_pgvector to improve naming accuracy
